power: Rerun AICL if charGing disable/enable
It is found that the USB settled ICL drop down when battery
charGing is disabled for full charge, but the ICL does not
adaptively increase for incremental system load or recharge.
In that case, it draws current from battery even a charger
is attached.
In order to recover from that situation, rerun AICL is an
effective way to recover the USB ICL.

Change-Id: Ic08691817bb51276187db558cdd5addd07fc85f0
Signed-off-by: Yonghua Yan <yanyh2@motorola.com>
Reviewed-on: https://gerrit.mot.com/1364085
SLTApproved: Slta Waiver
SME-Granted: SME Approvals Granted
Tested-by: Jira Key
Reviewed-by: Kenneth Kessler <kennykessler@motorola.com>
Submit-Approved: Jira Key