power: qpnp-smbcharger-mmi: Wait for FG Ready
IF the Max FG report -EAGAIN then it has not completed
the init of the battery profile and values could be
inaccurate.  So stop processing and reschedule heartbeat
if this is seen so a fresh 1st run of data can be taken.

Change-Id: I2070ba9e90338b7eb2ba8eb4f395739170810012
Signed-off-by: Kenny Kessler <kenny.kessler@motorola.com>
Reviewed-on: https://gerrit.mot.com/1376495
SME-Granted: SME Approvals Granted
SLTApproved: Slta Waiver
Tested-by: Jira Key
Reviewed-by: Cesar Augusto Marcelino dos Santos <cmsantos@motorola.com>
Reviewed-by: Ryan Lattrel <ryanl@motorola.com>
Reviewed-by: Kenneth Kessler <kennykessler@motorola.com>
Submit-Approved: Jira Key