ASoC: cs35l41: Prevent DSP wake when bypassed
Add an event handler to the PCM Source control
to set a state variable. Check this state variable
in the main amp event and don't send DSP mailbox
commands if the DSP is not routed to the ourput.

Cirrus-Jira-fixed: (CR)

Change-Id: I859f7e29949cf98fd39f4a02a6b3f64609785ce0
Signed-off-by: David Rhodes <david.rhodes@cirrus.com>
Reviewed-on: https://gerrit.mot.com/1504177
SME-Granted: SME Approvals Granted
SLTApproved: Slta Waiver
Tested-by: Jira Key
Reviewed-by: Wanlong Zhou <zhouwl@motorola.com>
Submit-Approved: Jira Key