mmi_info: Update ram info for ddr5
Update ram info to handle ddr5, which uses a new
struct from the bootloader with an extra flag.

Also re-work error handling, previously it was possible
that if a wrong size is returned, the pointer to the
smem struct might not be null and we will still try
to access it, causing a panic. So, do not add the
sysfs entry if it is not correct.

Change-Id: Ib195543175528d0477266281080c4dc4a96abd3c
Signed-off-by: Ryan Lattrel <ryanl@motorola.com>
Reviewed-on: https://gerrit.mot.com/1402645
SLTApproved: Slta Waiver
SME-Granted: SME Approvals Granted
Reviewed-by: Dong-Hua Yan <a18317@motorola.com>
Reviewed-by: Curtis Porter <porterc@motorola.com>
Tested-by: Jira Key
Reviewed-by: Konstantin Makariev <kmakariev@motorola.com>
Submit-Approved: Jira Key