ASoC: cs35l41: set default trigger as falling edge
There are irq storms observed on Racer DVT1 device. We have switched
the IRQ pin to GPIO56/GPIO57 from DVT1. The detailed reason is to be
investigated. Actually the irq pin is high when we set the irq trigger
as low level. If we do like this, we can set gpio-src-select to "0" in
dt for using this default falling edge.

Change-Id: Ie25da440037c5b9ac658b05e493d821cc84b951b
Signed-off-by: huangys <huangys@motorola.com>
Reviewed-on: https://gerrit.mot.com/1476897
SLTApproved: Slta Waiver
SME-Granted: SME Approvals Granted
Tested-by: Jira Key
Reviewed-by: Guobin Zhang <zhanggb@motorola.com>
Submit-Approved: Jira Key